Creating monomeric connection tool
