+++ src/schrodinger/sketcher/sketcher.qrc
@@ -46,6 +46,7 @@
     <file>icons/mode_erase.svg</file>
     <file>icons/mode_erase_dis.svg</file>
     <file>icons/mode_monomer.svg</file>
+    <file>icons/monomeric_connection.svg</file>
     <file>icons/periodic_table.svg</file>
     <file>icons/periodic_table_dis.svg</file>
     <file>icons/reaction_arrow.svg</file>
