+++ src/schrodinger/sketcher/ui/monomer_tool_widget.ui
@@ -7,7 +7,7 @@
     <x>0</x>
     <y>0</y>
     <width>100</width>
-    <height>235</height>
+    <height>274</height>
    </rect>
   </property>
   <property name="windowTitle">
@@ -1221,6 +1221,72 @@
      </widget>
     </widget>
    </item>
+   <item>
+    <widget class="Line" name="line">
+     <property name="orientation">
+      <enum>Qt::Horizontal</enum>
+     </property>
+    </widget>
+   </item>
+   <item>
+    <layout class="QHBoxLayout" name="horizontalLayout_6">
+     <item>
+      <widget class="QToolButton" name="covalent_or_disulfide_btn">
+       <property name="minimumSize">
+        <size>
+         <width>32</width>
+         <height>30</height>
+        </size>
+       </property>
+       <property name="maximumSize">
+        <size>
+         <width>32</width>
+         <height>32</height>
+        </size>
+       </property>
+       <property name="text">
+        <string>...</string>
+       </property>
+       <property name="icon">
+        <iconset resource="../sketcher.qrc">
+         <normaloff>:/icons/monomeric_connection.svg</normaloff>:/icons/monomeric_connection.svg</iconset>
+       </property>
+       <property name="checkable">
+        <bool>true</bool>
+       </property>
+       <attribute name="buttonGroup">
+        <string notr="true">monomeric_connection_group</string>
+       </attribute>
+      </widget>
+     </item>
+     <item>
+      <spacer name="horizontalSpacer">
+       <property name="orientation">
+        <enum>Qt::Horizontal</enum>
+       </property>
+       <property name="sizeHint" stdset="0">
+        <size>
+         <width>0</width>
+         <height>0</height>
+        </size>
+       </property>
+      </spacer>
+     </item>
+    </layout>
+   </item>
+   <item>
+    <spacer name="verticalSpacer_5">
+     <property name="orientation">
+      <enum>Qt::Vertical</enum>
+     </property>
+     <property name="sizeHint" stdset="0">
+      <size>
+       <width>0</width>
+       <height>0</height>
+      </size>
+     </property>
+    </spacer>
+   </item>
   </layout>
  </widget>
  <customwidgets>
@@ -1240,11 +1306,14 @@
    <header>schrodinger/sketcher/widget/amino_or_nucleic_toggle_button.h</header>
   </customwidget>
  </customwidgets>
- <resources/>
+ <resources>
+  <include location="../sketcher.qrc"/>
+ </resources>
  <connections/>
  <buttongroups>
   <buttongroup name="nucleic_monomer_group"/>
   <buttongroup name="amino_monomer_group"/>
   <buttongroup name="amino_or_nucleic_group"/>
+  <buttongroup name="monomeric_connection_group"/>
  </buttongroups>
 </ui>
